Add shift files into folders and adjust array fns
