+++ packages/common/src/utils/arrays/ArrayUtils.ts
@@ -1,0 +1,34 @@
+export const ArrayUtils = {
+  ifTheSameElseDefault: <T, K extends keyof T, J>(
+    someEntities: T[],
+    key: K,
+    defaultValue: J
+  ): J | T[K] => {
+    if (someEntities.length === 0) {
+      return defaultValue;
+    }
+    const entities = someEntities as [T, ...T[]];
+    const value = entities[0][key];
+    const allTheSame = entities.every(entity => {
+      return entity[key] === value;
+    });
+    return allTheSame ? value : defaultValue;
+  },
+  getUnitQuantity: (
+    arr: {
+      numberOfPacks: number;
+      packSize: number;
+    }[]
+  ) => {
+    return arr.reduce(
+      (sum, someEntity) => sum + someEntity.numberOfPacks * someEntity.packSize,
+      0
+    );
+  },
+  getSum: <T extends Record<K, number>, K extends string>(
+    arr: T[],
+    key: K
+  ): number => {
+    return arr.reduce((sum, someEntity) => sum + someEntity[key], 0);
+  },
+};
