+++ packages/common/src/utils/arrays/index.ts
@@ -1,6 +1,5 @@
-export * from './utils';
-export * from './reducers';
-export * from './sorters';
+export * from './ArrayUtils';
+export * from './SortUtils';
 
 import groupBy from 'lodash/groupBy';
 export { groupBy };
