+++ packages/requisitions/src/ResponseRequisition/api/hooks.ts
@@ -12,7 +12,7 @@ import {
   FieldSelectorControl,
   useFieldsSelector,
   SortController,
-  getColumnSorter,
+  SortUtils,
   useQueryParams,
   useMutation,
   useNotification,
@@ -117,7 +117,9 @@ export const useResponseLines = (): UseResponseLinesController => {
     const currentColumn = columns.find(({ key }) => key === sortBy.key);
     const { getSortValue } = currentColumn ?? {};
     return getSortValue
-      ? lines?.nodes.sort(getColumnSorter(getSortValue, !!sortBy.isDesc))
+      ? lines?.nodes.sort(
+          SortUtils.getColumnSorter(getSortValue, !!sortBy.isDesc)
+        )
       : lines?.nodes;
   }, [sortBy.key, sortBy.isDesc, lines]);
